pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tradeengine"
version = "0.1.0"
description = "本地加密貨幣交易引擎 - Local crypto trade engine with vectorBT"
requires-python = ">=3.9"
dependencies = [
    "vectorbt>=0.26",
    "pandas>=1.5",
    "numpy>=1.22,<2.0",
    "numba>=0.53",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.24",
    "websockets>=11.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "jinja2>=3.1",
    "typer>=0.9",
    "rich>=13.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "pyarrow>=12.0",
    "plotly>=5.15",
    "python-dotenv>=1.0",
    "eval-type-backport>=0.2.0",
    "asyncpg>=0.29",
    "cryptography>=41.0",
    "PyJWT[crypto]>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
tradeengine = "tradeengine.cli:app"

[tool.setuptools.packages.find]
include = ["tradeengine*"]

[tool.ruff]
line-length = 100
